Write the medium resolution arm as spectrograph 3
This is needed to avoid having to tell both the ETC and simulator that we
are using the medium resolution arm.  When you're running the simulator,
all other parameters (except number of exposures and input spectrum) are only
needed by the ETC, while the number of exposures and input spectrum are only
needed by the simulator.